Have gotten the switches fully working.
Added trigger events upon collision with player. Then if the player hits the interaction button the switch destroys the platform overhead which in turn drops the box for the player to move through the level.

Also have added a change color feature by changing the renderer component so when the switch is used it changes to a predefined color.