Triggers fully working
The guts of level 1 is now there, the triggers are fully working with each box and if the wrong box goes in the wrong goal the game automatically restarts. Have a new problem where the boxes get stuck on eachother when trying to push over the goal of another one. Might need to spawn a platform on top of a goal when the correct box is put in place to try and smoothen out the ground